+++ test/dummy/app/models/external_ref.rb
@@ -1,0 +1,4 @@
+class ExternalRef < ApplicationRecord
+  use_integer_primary_key
+  belongs_to :user
+end